requirements.txt
boto3==1.34.7
pyinstaller==6.3.0
rich==13.7.0
typer==0.9.0
PyGithub==2.1.1

